Move track_progress to helpers.utils
